security: translate OIDC claims to Biscuit facts dynamically, reorganize token logic, and add specifications references
