+++ cmd/sam-hub/main.go
@@ -39,8 +39,6 @@ import (
 	"time"
 
 	"github.com/biscuit-auth/biscuit-go/v2"
-	"github.com/biscuit-auth/biscuit-go/v2/datalog"
-	"github.com/biscuit-auth/biscuit-go/v2/parser"
 	"github.com/coreos/go-oidc/v3/oidc"
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/google/sam/api"
@@ -376,166 +374,6 @@ func (h *Hub) parseAndVerifyJWT(ctx context.Context, jwtStr string, allowedAudie
 	return claims, token, nil
 }
 
-func (h *Hub) mintBiscuitToken(claims jwt.MapClaims, token *oidc.IDToken, remotePeer peer.ID) ([]byte, error) {
-	var oidcRoles []string
-	if rolesAny, ok := claims["roles"].([]any); ok {
-		for _, r := range rolesAny {
-			if str, ok := r.(string); ok && str != "" {
-				oidcRoles = append(oidcRoles, str)
-			}
-		}
-	}
-
-	var oidcGroups []string
-	if groupsAny, ok := claims["groups"].([]any); ok {
-		for _, g := range groupsAny {
-			if str, ok := g.(string); ok && str != "" {
-				oidcGroups = append(oidcGroups, str)
-			}
-		}
-	}
-
-	oidcSub, _ := claims["sub"].(string)
-
-	// Resolve roles based on configured bindings and explicit OIDC roles
-	resolvedRoles := make(map[string]bool)
-	if h.Policy != nil {
-		// 1. Map OIDC groups and users to roles via configured bindings (RBAC mapping)
-		for _, b := range h.Policy.Bindings {
-			if b.Group != "" {
-				for _, cg := range oidcGroups {
-					if b.Group == cg {
-						resolvedRoles[b.Role] = true
-					}
-				}
-			}
-			if b.User != "" && oidcSub != "" {
-				if b.User == oidcSub {
-					resolvedRoles[b.Role] = true
-				}
-			}
-		}
-
-		// 2. Validate and accept pre-resolved OIDC roles directly if defined in policy (Zero-Trust check)
-		for _, r := range oidcRoles {
-			if _, exists := h.Policy.Roles[r]; exists {
-				resolvedRoles[r] = true
-			}
-		}
-	}
-
-	builder := biscuit.NewBuilder(h.KeyRing.GetCurrentKey())
-
-	if err := builder.AddAuthorityFact(biscuit.Fact{Predicate: biscuit.Predicate{
-		Name: api.FactExpiration,
-		IDs:  []biscuit.Term{biscuit.Date(token.Expiry)},
-	}}); err != nil {
-		return nil, fmt.Errorf("failed to add expiration fact: %w", err)
-	}
-
-	if err := builder.AddAuthorityFact(biscuit.Fact{Predicate: biscuit.Predicate{
-		Name: api.FactNode,
-		IDs:  []biscuit.Term{biscuit.String(remotePeer.String())},
-	}}); err != nil {
-		return nil, fmt.Errorf("failed to add node fact: %w", err)
-	}
-
-	if err := builder.AddAuthorityFact(biscuit.Fact{Predicate: biscuit.Predicate{
-		Name: api.FactClientPeerID,
-		IDs:  []biscuit.Term{biscuit.String(remotePeer.String())},
-	}}); err != nil {
-		return nil, fmt.Errorf("failed to add client_peer_id fact: %w", err)
-	}
-
-	if oidcSub != "" {
-		if err := builder.AddAuthorityFact(biscuit.Fact{Predicate: biscuit.Predicate{
-			Name: api.FactUser,
-			IDs:  []biscuit.Term{biscuit.String(oidcSub)},
-		}}); err != nil {
-			return nil, fmt.Errorf("failed to add user fact: %w", err)
-		}
-	}
-
-	// Assert original OIDC groups in the token (semantic audit trail)
-	seenGroups := make(map[string]bool)
-	for _, cg := range oidcGroups {
-		if seenGroups[cg] {
-			continue
-		}
-		seenGroups[cg] = true
-		if err := builder.AddAuthorityFact(biscuit.Fact{Predicate: biscuit.Predicate{
-			Name: api.FactGroup,
-			IDs:  []biscuit.Term{biscuit.String(cg)},
-		}}); err != nil {
-			return nil, fmt.Errorf("failed to add group fact: %w", err)
-		}
-	}
-
-	// Assert resolved authorized roles in the token
-	for role := range resolvedRoles {
-		if err := builder.AddAuthorityFact(biscuit.Fact{Predicate: biscuit.Predicate{
-			Name: api.FactRole,
-			IDs:  []biscuit.Term{biscuit.String(role)},
-		}}); err != nil {
-			return nil, fmt.Errorf("failed to add role fact: %w", err)
-		}
-
-		if h.Policy != nil {
-			if rolePolicy, ok := h.Policy.Roles[role]; ok {
-				for _, tool := range rolePolicy.MCP.AllowedServers {
-					if err := builder.AddAuthorityFact(biscuit.Fact{Predicate: biscuit.Predicate{
-						Name: api.FactMCPServer,
-						IDs:  []biscuit.Term{biscuit.String(tool)},
-					}}); err != nil {
-						logger.Errorw("Failed to add MCP tool fact to biscuit", "peer_id", remotePeer, "tool", tool, "error", err)
-					}
-				}
-				for _, target := range rolePolicy.Network.AllowedTargets {
-					if err := builder.AddAuthorityFact(biscuit.Fact{Predicate: biscuit.Predicate{
-						Name: api.FactNetworkTarget,
-						IDs:  []biscuit.Term{biscuit.String(target)},
-					}}); err != nil {
-						logger.Errorw("Failed to add network target fact to biscuit", "peer_id", remotePeer, "target", target, "error", err)
-					}
-				}
-				for _, customFact := range rolePolicy.CustomDatalog {
-					trimmed := strings.TrimRight(strings.TrimSpace(customFact), ";")
-					if trimmed == "" {
-						continue
-					}
-					func() {
-						defer func() {
-							if r := recover(); r != nil {
-								logger.Errorw("Panic parsing custom fact", "peer_id", remotePeer, "fact", trimmed, "recover", r)
-							}
-						}()
-						fact, err := parser.FromStringFact(trimmed)
-						if err != nil {
-							logger.Errorw("Failed to parse custom fact", "peer_id", remotePeer, "fact", trimmed, "error", err)
-							return
-						}
-						if err := builder.AddAuthorityFact(fact); err != nil {
-							logger.Errorw("Failed to add custom fact to biscuit", "peer_id", remotePeer, "fact", trimmed, "error", err)
-						}
-					}()
-				}
-			}
-		}
-	}
-
-	t, err := builder.Build()
-	if err != nil {
-		return nil, fmt.Errorf("failed to build biscuit: %w", err)
-	}
-
-	biscuitData, err := t.Serialize()
-	if err != nil {
-		return nil, fmt.Errorf("failed to serialize biscuit: %w", err)
-	}
-
-	return biscuitData, nil
-}
-
 // startWatchdog periodically checks for peers that have connected but not completed OIDC
 // authentication within the grace period, and evicts them from the network.
 
@@ -850,57 +688,6 @@ func isLoopbackOrLinkLocal(addr multiaddr.Multiaddr) bool {
 		}
 	}
 	return false
-}
-
-func (h *Hub) verifyBiscuit(biscuitData []byte, remotePeer peer.ID) (*biscuit.Biscuit, error) {
-	b, err := biscuit.Unmarshal(biscuitData)
-	if err != nil {
-		return nil, fmt.Errorf("malformed biscuit: %w", err)
-	}
-
-	var authOpts []biscuit.AuthorizerOption
-	if h.BiscuitTimeout > 0 {
-		authOpts = append(authOpts, biscuit.WithWorldOptions(datalog.WithMaxDuration(h.BiscuitTimeout)))
-	}
-
-	keys := h.KeyRing.GetAllValidPublicKeys()
-	var lastErr error
-	for _, pubKey := range keys {
-		authorizer, err := b.Authorizer(pubKey, authOpts...)
-		if err != nil {
-			lastErr = err
-			continue
-		}
-
-		authorizer.AddFact(biscuit.Fact{
-			Predicate: biscuit.Predicate{
-				Name: "time",
-				IDs:  []biscuit.Term{biscuit.Date(time.Now())},
-			},
-		})
-
-		timeCheck, err := parser.FromStringCheck(`check if time($time), expiration($exp), $time <= $exp`)
-		if err != nil {
-			lastErr = err
-			continue
-		}
-		authorizer.AddCheck(timeCheck)
-
-		rule, err := parser.FromStringPolicy("allow if true")
-		if err != nil {
-			lastErr = err
-			continue
-		}
-		authorizer.AddPolicy(rule)
-
-		if err := authorizer.Authorize(); err == nil {
-			return b, nil
-		} else {
-			lastErr = err
-		}
-	}
-
-	return nil, fmt.Errorf("no valid key found for verification: %v", lastErr)
 }
 
 func (h *Hub) HandleAuthHandshake(s network.Stream) {
