+++ site/content/docs/development/policy.md
@@ -12,10 +12,57 @@ The Hub automatically translates OIDC claims into undeniable cryptographic facts
 | :--- | :--- | :--- |
 | `sub` | `user("<sub-id>")` | The unique subject ID from the identity provider. |
 | `email` | `email("<email>")` | The user's email address (if present). |
-| `groups` / `roles` | `role("<role-name>")` | One fact is injected for *each* role/group the user possesses. |
-| Generated FQDN | `name("<fqdn>")` | The collision-proof mesh name granted by the Hub. |
-| Peer ID | `node("<peer-id>")` | Binds the token to the specific agent's libp2p cryptographic identity. |
-| Expiration | `time(<date>)` | The token expiration date based on the OIDC session. |
+| `groups` | `group("<group-name>")` | One fact is injected for *each* group the user possesses. |
+| `roles` / Resolved Roles | `role("<role-name>")` | One fact is injected for *each* role mapped or direct role. |
+| Peer ID | `node("<peer-id>")`, `client_peer_id("<peer-id>")` | Binds the token to the specific agent's libp2p cryptographic identity. |
+| Expiration | `expiration(<date>)` | The token expiration date based on the OIDC session. |
+
+### 1.1 Translating Identity to Capability: OIDC to Biscuit
+The core innovation of the SAM Network's security model is translating standard web identity (OIDC JSON Web Tokens defined in [OpenID Connect Core 1.0](https://openid.net/specs/openid-connect-core-1_0.html#Claims)) into decentralized capability tokens (Biscuits). This translation happens securely at the `sam-hub` during the authentication phase.
+
+The Datalog authority facts generated by the Hub are constructed using the Biscuit [Symbol Table specification](https://doc.biscuitsec.org/reference/specifications.html#symbol-table), ensuring compact serialization and cryptographically undeniable credentials.
+
+Here is exactly how an OIDC JWT is transformed into a policy-ready Biscuit:
+
+#### 1. Claim Extraction
+When an agent or user submits a valid OIDC JWT, the sam-hub verifies the token's signature against the Identity Provider. Once validated, the Hub extracts the payload claims—typically attributes like sub (subject), email, and custom arrays like groups or roles.
+
+#### 2. Datalog Fact Generation
+Biscuit policies are written in Datalog, a declarative logic language. The sam-hub acts as a translator, mapping the JSON claims from the OIDC token into immutable Datalog facts.
+
+For example, an incoming OIDC payload like this:
+
+```json
+{
+  "sub": "user-12345",
+  "email": "agent@google.com",
+  "groups": ["beta-testers", "engineering"]
+}
+```
+
+Is translated by the Hub into the following Datalog facts:
+
+```datalog
+user("user-12345");
+email("agent@google.com");
+group("beta-testers");
+group("engineering");
+```
+
+#### 3. Minting the Authority Block
+The sam-hub takes these generated Datalog facts and embeds them into the Authority Block of a brand new Biscuit token. The Hub signs this block with its private cryptographic key.
+
+Because the facts are sealed in the Authority Block by the Hub, any sam-node in the mesh can implicitly trust that the user holding the Biscuit possesses those specific emails and groups.
+
+#### 4. Policy Evaluation at the Node
+When the agent presents the Biscuit to a sam-node to execute a tool, the node evaluates its local policies against the facts embedded in the token.
+
+Because the OIDC claims were translated into Datalog, a node administrator can write elegant, logic-based rules in policy.go or their YAML configs:
+
+```datalog
+// The node will only execute the tool if the Hub certified the agent is in the engineering group
+allow if group("engineering");
+```
 
 ## 2. Hub Policy Schema (`policies.yaml`)
 Admins define central permissions by mapping OIDC roles to specific capabilities.
